+++ runtime/admin/src/main/java/org/apache/polaris/admintool/PolarisAdminTool.java
@@ -31,7 +31,6 @@ import picocli.CommandLine.HelpCommand;
     subcommands = {
       HelpCommand.class,
       BootstrapCommand.class,
-      BootstrapMetricsCommand.class,
       PurgeCommand.class,
       NoSqlCommand.class,
     })